+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * update dashboard widget for MarComm publishing of posts
 * update to correctly identify Opera browser in `Report a Bug`
+* remove unused item from **Extra Settings** tab
 
 #### 3.4.1 / 2023-05-26
 * **Report a Bug** only for logged in users